backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.38
python-jose==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiosqlite==0.20.0
pydantic==2.10.6
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1